プロジェクト構成の変更

+++ Othello/AgentStrategy/RandomBot.cpp
@@ -14,7 +14,7 @@ Point RandomBot::answer(const Othello& othello) {
 
 	std::random_device rnd;
 	std::mt19937 mt(rnd());
-	std::uniform_int_distribution<> randSelect(0, canPutPoints.size()-1);
+	std::uniform_int_distribution<> randSelect(0, (int)canPutPoints.size()-1);
 	const int randIndex = randSelect(mt);
 
 	return canPutPoints[randIndex];
